Move WantedAppiumCapabilities class
